Tha average daily drive time is now an input

+++ CSVhandling/DataGenerator.py
@@ -2,7 +2,7 @@ import random
 
 
 class DataGenerator:
-    def __init__(self, season: str, avgDailyDrive: int, initialPressures: dict):
+    def __init__(self, season: str, avgDailyDrive: float, initialPressures: dict):
         self.__season = "ALL"
         if season.upper() in ["SUMMER", "WINTER"]:
             self.__season = season.upper()
